gsctool: Add host side bid check
Even though the GSC RW FW will validate the bid constraints before
attempting a FW update, the gsctool side can validate if the potential
GSC FW will load on the device or not. This is especially helpful for NT
images as the board id constraints are at the end of the image, and the
GSC FW cannot validate the constraints until the entire image has been
loaded into GSC flash.

To give end-users a better experience, gsctool now fails early if the
board id constraints will prevent this image from running on the
connected GSC device.

Also, added the `--skip_bid_check` command line option to override this
check and send the firmware to the GSC device anyway. The GSC device
should error when trying to write this potential firmware or finalize
the potential firmware.

BUG=none
TEST=See that bid constraint firmware is getting blocked when it should
TEST=See that adding `--skip_bid_check` allow for "bad" FW to be send to
     GSC device and that GSC eventually send back an error
TEST=See that normal images are not getting blocked by new check.
TEST=BED=DT make tast TAST_EXPR=gscdevboard.GSCFactoryUpdate.ti50_0_21_1

Change-Id: I8e4007f3fde95f5109693df52be0676ec33b72bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6618735
Tested-by: Jett Rink <jettrink@google.com>
Reviewed-by: Vadim Bendebury <vbendeb@google.com>
Commit-Queue: Jett Rink <jettrink@google.com>